requirements.txt
streamlit
wikipedia-api
wikipedia
google-generativeai
textblob
dotenv
matplotlib
feedparser